Move gitlab project to gitlab/pj
