+++ src/remote/query.rs
@@ -10,7 +10,8 @@ use crate::{
         GithubMemberFields, GithubMergeRequestFields, GithubPipelineFields, GithubProjectFields,
     },
     gitlab::mr::GitlabMergeRequestFields,
-    gitlab::{GitlabMemberFields, GitlabPipelineFields, GitlabProjectFields},
+    gitlab::pj::{GitlabMemberFields, GitlabProjectFields},
+    gitlab::GitlabPipelineFields,
     http::{self, Body, Headers, Paginator, Request, Resource},
     io::{HttpRunner, Response},
     json_load_page, json_loads,
